+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Coding standard.
 - Change log.
+- Event hydration system.
 
 ### Fixed
 
@@ -18,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Event storage exceptions.
-- Bumped PHP version requirement to `7.2`.
+- Bumped PHP min version requirement to `7.2`.
 - Updated dependencies.
+- Event type to a static method.